+++ src/test/status.test.ts
@@ -330,3 +330,76 @@ test('publishing action removes other statuses', async () => {
   })
   test.is(published.status, 'published')
 })
+
+test('unique path check in unpublished', async () => {
+  const data = {title: 'John Doe', path: 'john-doe'}
+  const db = new LocalDB(cms.config)
+  const unpublished = await db.create({
+    type: Page,
+    set: {path: 'page1', title: 'Unublished'},
+    status: 'draft'
+  })
+  const child1 = await db.create({
+    type: Page,
+    parentId: unpublished._id,
+    set: data,
+    status: 'draft'
+  })
+  const child2 = await db.create({
+    type: Page,
+    parentId: unpublished._id,
+    set: data,
+    status: 'draft'
+  })
+  await db.publish({
+    id: unpublished._id,
+    status: 'draft'
+  })
+  // This should not overwrite child1 which has the same path
+  await db.publish({
+    id: child2._id,
+    status: 'draft'
+  })
+  const first = await db.first({id: child1._id, select: Entry, status: 'draft'})
+  test.ok(first)
+})
+
+test("it's possible to archive unpublished entries", async () => {
+  const db = new LocalDB(cms.config)
+  const unpublished = await db.create({
+    type: Page,
+    set: {path: 'page1', title: 'Unpublished'},
+    status: 'draft'
+  })
+  const child1 = await db.create({
+    type: Page,
+    parentId: unpublished._id,
+    set: {path: 'child1', title: 'Child 1'},
+    status: 'draft'
+  })
+  const child2 = await db.create({
+    type: Page,
+    parentId: unpublished._id,
+    set: {path: 'child2', title: 'Child 2'},
+    status: 'draft'
+  })
+  await db.archive({id: unpublished._id})
+  const archivedUnpublished = await db.get({
+    id: unpublished._id,
+    select: Entry,
+    status: 'archived'
+  })
+  test.is(archivedUnpublished.status, 'archived')
+  const archivedChild1 = await db.get({
+    id: child1._id,
+    select: Entry,
+    status: 'archived'
+  })
+  test.is(archivedChild1.status, 'archived')
+  const archivedChild2 = await db.get({
+    id: child2._id,
+    select: Entry,
+    status: 'archived'
+  })
+  test.is(archivedChild2.status, 'archived')
+})